delete old footer
