+++ src/components/footer/index.js
@@ -1,71 +1,328 @@
-import React from 'react';
+import React, { Component } from 'react'
+import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
 
-import RightFooter          from './RightFooter';
-import MidFooter            from './MidFooter';
-import LeftFooter           from './LeftFooter';
+import * as routes from '../../constants/routes';
 
-const Footer = () => {
-    return (
-        <>
-            <LeftFooterGrid className="grid-footer-left">
-                <LeftFooter/>
-            </LeftFooterGrid>
-            <MidFooterGrid className="grid-footer-mid">
-                <MidFooter/>
-            </MidFooterGrid>
-            <RightFooterGrid className="grid-footer-right">
-                <RightFooter/>
-            </RightFooterGrid>
-        </>
-    )
+import LeviEikoBrand from '../../levieiko';
+
+export default class Footer extends Component {
+
+    render(){
+        const { toggleEmailSignup, scrollToTop } = this.props;
+        return (
+            <FooterWrapper className="grid-footer">
+                <QuickLinksSection>
+                    <h1>QUICK links</h1>
+                    <div>
+                        <section>
+                            <NavLink activeClassName="nav-active" to={routes.HOME} onClick={scrollToTop}>HOME</NavLink>
+                            <NavLink activeClassName="nav-active" to={routes.EVNT} onClick={scrollToTop}>SCHEDULE</NavLink>
+                            <NavLink activeClassName="nav-active" to={routes.INFO} onClick={scrollToTop}>ABOUT US</NavLink>
+                            <button className="link-btn"
+                                name="contactForm" 
+                                value="contactRequest" 
+                                onClick={(e) => {toggleEmailSignup(e)}}
+                                >
+                                CONTACT US
+                            </button>
+                        </section>
+                        <section>
+                            <NavLink activeClassName="nav-active" to={routes.SERV} onClick={scrollToTop}>WELLNESS SERVICES</NavLink>
+                            <div>
+                                <NavLink activeClassName="nav-active" to={routes.WELL} onClick={scrollToTop}>PERSONAL WELLNESS</NavLink>
+                                <NavLink activeClassName="nav-active" to={routes.CORP} onClick={scrollToTop}>CORPORATE WELLNESS</NavLink>
+                                <NavLink activeClassName="nav-active" to={routes.YOGA} onClick={scrollToTop}>YOGA THERAPY</NavLink>
+                                <NavLink activeClassName="nav-active" to={routes.PREG} onClick={scrollToTop}>MATERNAL HEALTH</NavLink>
+                                <NavLink activeClassName="nav-active" to={routes.COAC} onClick={scrollToTop}>SPEAKING & COACHING</NavLink>
+                            </div>
+                        </section>
+                    </div>
+                    <div>
+                        <LeviEikoBrand/>
+                    </div>
+                </QuickLinksSection>
+                <LogoSection>
+                    <div>
+                        <img src="../icon_white.png" alt="pranayama yoga breath work" />
+                        <Prana className="pranaText">PRANA</Prana>
+                        <Wellness className="wellnessText">wellness</Wellness>
+                    </div>
+                </LogoSection>
+                <ContactSection>
+                    <h1>HAVE QUESTIONs?</h1>
+                    <div>
+                        <h3>Stephanie@pranawellnessworks.com</h3> 
+                        <button 
+                            className="purple"            
+                            name="contactForm" 
+                            value="contactRequest" 
+                            onClick={(e) => {toggleEmailSignup(e)}
+                            }>
+                            SEND US A MESSAGE
+                        </button>
+                    </div>
+                    <div>
+                        <LeviEikoBrand/>
+                    </div>
+                    <div>
+                        <a href="https://www.travelslay.co/" target="_blank" rel="noopener noreferrer">
+                            <img src="../travel_slay_white.png" alt="Travel Slay"/>
+                        </a>
+                        <section>explore fitness and travel</section>
+                    </div>
+                </ContactSection>
+            </FooterWrapper>
+        )
+    }
 }
-const FooterGrid = styled.div`
-  background-color: var(--purple);
-  padding-top: 2vw;
-  position: relative;
-`;
 
-const LeftFooterGrid = styled(FooterGrid)`
-  grid-area: footer-left;
-  @media screen and (max-width: 945px) {
-    text-align: center;
-    div {
-      margin-left: auto;
-      margin-right: auto;
+
+const ContactSection = styled.section`
+    position: relative;
+    grid-area: contact;
+    > div {
+        text-align: center;
+        &:nth-of-type(2) {
+            visibility: hidden;
+            position absolute;
+            bottom: 5vw;
+            left: 5vw;
+        }
+        &:last-of-type {
+            position: absolute;
+            bottom: 2.5vw;
+            right: 2.5vw;
+            width: 15vw;
+            text-align: center;
+            > section {
+                color: #fff;
+                font-size: 1vw;
+            }
+            > a {
+                > img {
+                    width: 100%;
+                    margin-bottom: -2vw;
+                }
+            }
+        }
+
+        > h3 {
+            text-align: center;
+            text-transform: uppercase;
+            color: var(--lightpurple);
+            font-size: 1.4vw;
+            font-weight: 100;
+        }
+
+        > button {
+            margin-top: 2vw;
+        }
+    }
+    @media screen and (max-width: 945px) {
+        > div {
+            &:first-of-type {
+                padding-bottom: 40vw;
+            }
+            &:nth-of-type(2) {
+                visibility: visible;
+            }
+            &:last-of-type {
+                position: absolute;
+                bottom: 5vw;
+                right: 5vw;
+                width: 35vw;
+                text-align: center;
+                > section {
+                    color: #fff;
+                    font-size: 2.5vw;
+                }
+                > a {
+                    > img {
+                        width: 100%;
+                        margin-bottom: -2vw;
+                    }
+                }
+            }
+    
+            > h3 {
+                font-size: 4vw;
+                margin-bottom: 5vw;
+            }
+    
+            > button {
+                margin-top: 2vw;
+            }
+        }
     }
 `;
 
-const MidFooterGrid = styled(FooterGrid)`
-  grid-area: footer-mid;
+const LogoSection = styled.section`
+    grid-area: logo;
+    > div {
+        text-align: center;
+
+        > img {
+            width: 100%;
+            margin: 0 auto -2vw;
+        }
+    }
+    @media screen and (max-width: 945px) {
+        > div {
+            > img {
+                width: 40%;
+                margin: 0 auto -2vw;
+            }
+        }
+    }
+`;
+const QuickLinksSection = styled.section`
+    grid-area: links;
+    position: relative;
+    .link-btn {
+        background: transparent;
+        border: none;
+        padding: 0;
+        margin: 0;
+    }
+    > div {
+        &:first-of-type {      
+            a {
+                display: block;
+                text-decoration: none;
+            }
+            > section {
+                display: inline-block;
+                vertical-align: top;
+                
+                button, 
+                > a {
+                    color: #fff;
+                    font-size: 1.6vw;
+                    padding-left: 1vw;
+                    margin-bottom: 1vw;
+                    letter-spacing: .05vw;
+                }
+                &:first-of-type {                    
+                }
+                &:last-of-type {                    
+                    > div {
+                        padding-left: 2vw;
+                        font-size: 1.25vw;
+                        > a {
+                            color: var(--lightpurple)!important;
+                            margin-bottom: .8vw;
+                            letter-spacing: .025vw;
+                        }
+                    }
+                }
+            }
+        }
+        &:last-of-type {
+            position: absolute;
+            bottom: 2.5vw;
+            left: 2.5vw;
+        }
+    }
+    @media screen and (max-width: 945px) {
+
+        > div {
+            &:first-of-type {      
+                a {
+                    display: block;
+                    text-decoration: none;
+                }
+                > section {
+                    display: inline-block;
+                    vertical-align: top;
+                    
+                    button, 
+                    > a {
+                        font-size: 6vw;
+                        padding-left: 5vw;
+                        margin-bottom: 2.5vw;
+                        letter-spacing: .05vw;
+                    }
+                    &:first-of-type {                    
+                    }
+                    &:last-of-type {                    
+                        > div {
+                            padding-left: 10vw;
+                            font-size: 5.25vw;
+                            > a {
+                                margin-bottom: 2vw;
+                                letter-spacing: .025vw;
+                            }
+                        }
+                    }
+                }
+            }
+            &:last-of-type {
+                visibility: hidden;
+            }
+        }
+    }
+`;
+const Prana = styled.section`
+  text-transform: uppercase;
+  font-weight: 100;
+  color: #fff;
+  font-size: 3.4vw;
+  margin: -1vw auto;
+  @media screen and (max-width: 945px) {
+    font-size: 5vw;
+    padding-left: .5vw;
+  }
+`;
+
+const Wellness = styled.section`
   text-align: center;
-  // section {
-  //   margin: 3vw auto 1vw auto;
-  //   font-style: normal;
-  //   font-weight: normal;
-  //   font-size: 1.2vw;
-  //   letter-spacing: .3vw;
-  //   text-transform: uppercase;
-  //   color: var(--lightgreen);
-  //   text-align: center;
-  // }
-  // @media screen and (max-width: 945px) {
-  //   section {
-  //     margin: 3vw auto 1vw auto;
-  //     font-weight: bold;
-  //     font-size: 2.8vw;
-  //     letter-spacing: 3px;
-  //   }
-  //   input {
-  //     width: 90vw;
-  //     height: 10vw;
-  //     font-size: 5vw;
-  //   }
-  // }
+  color: #FFF;
+  font-family: 'Merriweather', sans-serif;
+  font-style: italic;
+  text-transform: lowercase;
+  font-size: 3vw;
+  @media screen and (max-width: 945px) {
+    font-size: 3.5vw;
+    margin-top: -6px;
+  }
 `;
 
-const RightFooterGrid = styled(FooterGrid)`
-  grid-area: footer-right;
-`;
+const FooterWrapper = styled.div`
+    background-color: var(--purple);
+    grid-area: footer;
+    width: 100%;
+    height: 100%;
+    min-height: 400px;
 
-export default Footer;
+    display: grid;
+    grid-template-areas: 'links logo contact';
+    grid-template-columns: 5fr 3fr 5fr; 
+
+    > section {
+        > h1 {
+            color: var(--lightgreen);
+            text-align: left;
+            border-bottom: 1px solid var(--lightgreen);
+            padding: 0;
+            font-size: 2.6vw;
+            letter-spacing: .1vw;
+            font-weight: 100;
+            margin-bottom: 1.5vw;
+        }
+        padding: 2.5vw;
+    }
+    @media screen and (max-width: 945px) {
+        grid-template-areas: 'logo' 'links' 'contact';
+        grid-template-columns: 100vw; 
+        grid-template-rows: auto auto auto;
+        > section {
+            > h1 {
+                font-size: 8vw;
+                letter-spacing: .3vw;
+                margin-bottom: 5vw;
+            }
+            padding: 5vw;
+        }
+    }
+`;
